refactor(vscode): use extension host i18n for autocomplete
