+++ packages/kilo-vscode/src/services/autocomplete/AutocompleteServiceManager.ts
@@ -1,6 +1,6 @@
 import crypto from "crypto"
 import * as vscode from "vscode"
-import { t } from "./shims/i18n"
+import { t } from "../i18n"
 import { TelemetryProxy, TelemetryEventName } from "../telemetry"
 import { AutocompleteStatusBar } from "./AutocompleteStatusBar"
 import { AutocompleteCodeActionProvider } from "./AutocompleteCodeActionProvider"
